master

+++ db/migration/000001_init_schema.up.sql
@@ -1,18 +1,18 @@
 CREATE TABLE `accounts` (
   `id` int PRIMARY KEY AUTO_INCREMENT,
   `owner` varchar(255) NOT NULL,
-  `balance` bigint NOT NULL,
+  `balance` int NOT NULL,
   `currency` varchar(255) NOT NULL,
-  `created_at` bigint NOT NULL,
-  `updated_at` bigint NOT NULL,
+  `created_at` int NOT NULL,
+  `updated_at` int NOT NULL,
   `timezone` varchar(255)
 );
 
 CREATE TABLE `entries` (
   `id` int PRIMARY KEY AUTO_INCREMENT,
   `account_id` int,
-  `amount` bigint NOT NULL COMMENT 'can be nagative',
-  `created_at` bigint NOT NULL,
+  `amount` int NOT NULL COMMENT 'can be nagative',
+  `created_at` int NOT NULL,
   FOREIGN KEY (`account_id`) REFERENCES `accounts` (`id`) ON UPDATE CASCADE ON DELETE CASCADE
 );
 
@@ -20,8 +20,8 @@ CREATE TABLE `transfers` (
   `id` int PRIMARY KEY AUTO_INCREMENT,
   `from_account_id` int,
   `to_account_id` int,
-  `amount` bigint NOT NULL COMMENT 'must be positive',
-  `created_at` bigint NOT NULL,
+  `amount` int NOT NULL COMMENT 'must be positive',
+  `created_at` int NOT NULL,
   FOREIGN KEY (`from_account_id`) REFERENCES `accounts` (`id`) ON UPDATE CASCADE ON DELETE CASCADE,
   FOREIGN KEY (`to_account_id`) REFERENCES `accounts` (`id`) ON UPDATE CASCADE ON DELETE CASCADE
 );
